Have Error In FlightLog

+++ src/main/java/org/mdt/ulsanproject/repository/FlightLogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FlightLogRepository extends JpaRepository<FlightLog, Integer> {
-
+public interface FlightLogRepository extends JpaRepository<FlightLog, Long> {
+    // Custom query methods can be added here if necessary
 }
